Renamed mountain range visualization file.
No changes made to the file other than to chaange the name.  However,
it is important to do some cleanup on the implementation when there is a
chance, as it is currently rather messy.  Also, it might be nice to add
the ability to do a quick image test of the mountain visualization
capability.  At the moment, it is testing the file with a matplotlib
plot which, though, providing reasonable information, is visually not
particularly appealing.